chore: 역할별 관심사 분리를 위한 코드 구조 정리
